fix #356 (change callsign dialog steals focus)
#   11-23-17   TMG       fix #356 (change callsign dialog should not pop
up until
#                         its new entry widget is active (i.e. the
active stack item)

Note, it's debatable as to which is more intuitive: pop up the change
callsign dialog immediately, and steal focus, as it was before this bug
fix - maybe add a bell or such to get the user's attention - or to go
ahead with this bug fix.  Time will tell which seems to work out better.