work
When sending to a specific device, use the com port on which that device was most recently heard from; keep track of this as an additional element in self.fsLog.
Still need to get it to automatically try the other port after failed response.
Note: if data is received on both com ports within the same fsCheck interval, and fsBuffer has data from both com ports, then the device's 'latest' com port could be saved incorrectly.  This should be extremely rare and wouldn't really be a big issue - the operator would just have to wait a few other seconds for it to try the other port.