WIP - significant timing cleanup
cleaned up handling of hold time and continue time concepts in various places in the code; need to test more before merging
NOTE that resetLastModAge will be called on every new incoming fs/nx call, even if there is no GPS data, due to radioLocField.setText in fsParse.  Not sure if this is appropriate.  Could easily be changed if needed.