WIP entry-hidden popup
first version of entry-hidden popup: when ActivationChange event is called on any of the newEntry-related windows (newEntryWindow, callsign change dialog, clue dialog, subject located dialog - need to also do this for non-radio clue dialog), show the non-modal stays-on-top newEntryWindowHiddenPopup - and close it when an important window is activated.  Need to test this more thoroughly (and add to non-radio clue dialog).