WIP
blink the pending-entry dialog, but make it silent (no beeps).  This seems more appropriate since it may be pretty often than the pending entry loses focus, on purpose, when the operator does something in a different program (web browser/caltopo, word, excel, etc).

This portion of #683 is ready for test and get feedback from operators; but remember that this portion is just one item ('add a prominent indication ...') on the list of items for #683.