WIP automatic entry
cleaned up timing and label text; discovered that it may not make sense to delete all text from the message body when subject located dialog is automatically opened based on keywords, in case the typed message contained important things before the keywords that should show up in the final saved entry, and not just in the automatically-opened entry.  Related to #676.  Should address that in another commit (on this branch is OK), if at all.